backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
geoalchemy2==0.15.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Geospatial
geopandas==0.14.4
shapely==2.0.4
rasterio==1.3.10
pyproj==3.6.1
fiona==1.9.6

# Google Earth Engine
earthengine-api==0.1.401

# Scheduler
apscheduler==3.10.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0

# Excel/CSV
openpyxl==3.1.2
pandas==2.2.2
